attack and release are exponential
envtable is now an array of time constants used to change the rate of
exponential rise/decay. the env.py script spits out the appropriate
values for the user's choice of 256 durations in seconds (i.e. the knob
gradations.)

the attack/release knob scale is piecewise; the first "half" (not
exactly 128 entries because the release knob's ADC channel is
nonlinear) is a linear scale from instantaneous to 1 second, and the
second "half" is an exponential scale from 1 second to 10 seconds.