eliminated "click" when a key is pressed
this is a big change, and unfortunately a sad one. we can no longer use
PWM to control the envelope volume, because the DAC output isn't
bipolar. a signal level of 0 corresponds to a DAC value of 0x800, a dc
offset of Vref/2. if the attack is instantaneous, playing a note
immediately causes the DAC output to jump from 0x000 to 0x800, causing
an audible click. it's not noticeable with a sharp waveform like a pure
pulse wave, but with a filtered wave it's quite unfortunate.

the solution is to downscale the output signal by the envelope value in
the interrupt handler. our total output precision is now 12 bits, which
is kinda sad.

on the plus side, we'll no longer need to use a PWM output, and some
components can be removed. the DAC's Vref can be connected to a pot for
volume control, and the newly-freed I/O line can be used for another
control, or maybe to disable MIDI input during programming mode.